+++ src/Engines/AlgoliaEngine.php
@@ -100,4 +100,20 @@ class AlgoliaEngine extends BaseAlgoliaEngine
 
         $index->clearObjects();
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function filters(Builder $builder): array
+    {
+        $operators = ['<', '<=', '=', '!=', '>=', '>'];
+
+        return collect($builder->wheres)->map(function ($value, $key) use ($operators) {
+            if (ends_with($key, $operators) || starts_with($value, $operators)) {
+                return $key.' '.$value;
+            }
+
+            return $key.'='.$value;
+        })->values()->all();
+    }
 }